Fix tooltip display on patch review (and probably other places as well)
Don't turn uint -> int, but the other way around. Otherwise, we were setting
std::max<uint> on the distance, which was then turned to int becoming a
negative number, which broke the premise of the comparison.
Instead, turn an int that we know is positive (addition of two abs() calls)
into a uint.