Use oldcpp behavior when inserting include paths.
This is much better, in my opinion. It allows one to easily include
long paths as well as change existing includes more easily as the
rest of the line gets removed as well.

Additionally, comments get taken into account as well as whitespaces.
An extensive unit test suite is added.

REVIEW: 122169