Make code completion faster
Instead of calling findDeclarations for each item found by clang, call
allDeclarations once and cache the result. In case it can't find
anything fall back to the findDeclarations.

This decreases code-completion time from ~2000ms to ~200ms for me in the
worst case.

As a drawback the code-completion time can increase in some cases
(e.g. for member access completion from ~60ms to ~200ms), but it's still
bearable.

REVIEW: 123810