Optimize findForwardDeclaration
The findDeclarations call is very time consuming, so
don't use it when not needed. Instead check if the CXType can
actually be a forward declaration, if it's not the case return early.

This greatly decreases time spend on parsing plain C files
(for me from 40 to 10 seconds in some cases).