Make sense out of the patch review initialization process
So far, it was just taking any review_* area that wasn't used elsewhere.
This created the following workflow: area switch -> closeViews from previous
area -> restore new Area -> close views -> create views for the patch.
Hence the huge flickering.
What we do now is to just clear the review working set before applying it,
so we ensure we restore an empty document set. Then we can just populate it
with what we want, however we want.