Introduction of a Concentration mode
One of the perceived problems in KDevelop that affect me the most is that
we're visually heavy. And we are. This patch doesn't address this, but I
believe it will start to give us some hints on where we want to go.

This patch, basically adds a new togglable state to the main window, which
allows our user to reduce some of the noise in the UI, allowing him to
concentrate on what he came to do. This has a big advantage: we get to
assume that the user is familiar with the shortcuts.

What it does:
* It hides all the toolview handles (not the toolviews).
* It hides the menu bar. Since I realized I was using Alt+* to access some
menus, it's showing the menu when Alt is pressed.
* It adds a toolbar that lets the user get out of the concentration mode,
and also receives the menu widget (with the working sets and area button).
* Also since I realized I was using some other shortcuts, I took the
liberty to adapt them, I think they're far easier to remember than the
previous ones.
Replaces the container tab bar to work without it (it has a label instead,
long story, I know).

REVIEW: 123878