Start integration of KTextEditor::Plugin into the KDevelop framework.
We wrap KTextEditor::Plugin in KDevelop::IPlugin. This works quite
nicely, except for the toolview factory concept we have in KDevelop.
To workaround the KTE API limitations we prevent the widget we
return via MainWindow::createToolView from ever getting deleted.
Only on shutdown, or when the factory gets killed, we actually destroy
the widget. The surrounding widget still gets deleted from KDevelop.

Note that for now only the snippets plugin is loaded and displayed.
We do not yet implement most of the KTE plugin API, and will need
to add more features in the future. This should lay the foundation
for future work in that direction though. We will want:

- support the MainWindow::viewBar concept to share a search bar
across multiple split views
- add the split / quit API for better VI mode integration
- enable more plugins and add whatever API they depend on. We will
want to kill our plugins and replace them with Kate ones for:
  - open documents (if it supports non-KTE documents)
  - search & replace
  - filesystem (if it allows opening of non-text files)
  - sql, ...?