Added look-ahead matching code completion
Known shortcomings:

1)Auto type (and many other types) is not exposed through LibClang. So
we assign DelayedType to it instead of IdentifiedType, therefore no
declaration attached to the type, hence no look-ahead completion.

2)LibClang missing API to determine expected code completion type. To
workaround it, now we use types of declarations from the best mathes
group. Therefore if no declarations found with a high enough priority,
no items will be added to the look-ahead completion group.

As a result the look-ahead completion works only in the simpliest cases.

REVIEW: 124601