qmljs: Avoid infinite recursion when the declaration of a structure instance points to itself
Such declarations are difficult to produce but may arise in snippets like this
(see unit test added by this commit):

function () {
    this.i_dont_exist = new this.i_dont_exist()
}

This line confuses new (that creates a structure) and the equal sign (that assigns
contexts and merges types in order to support type inference). The result is that
i_dont_exist has a structure type whose declaration is "i_dont_exist" itself.

BUG: 352122