Fix '.' to '->' member access replacement
Clang doesn't provide diagnostic for it. So let't try to
manually determine whether this is the case, this is not so hard
after all.

This also means that now we use clang_codeCompleteAt twice, but
it doesn't matter much here as the input is already invalid,
hence no completion items.

REVIEW: 125305