Correctly remove stale next bucket links when items are removed
This fixes an assertion that can occur in an edge case when there
is a clashing monster bucket followed by another clashing item.

This should improve index/findIndex performance as they no longer
need to search these stale links. It's hard to predict if this
is significant, but at least it should prevent performance
degradation in item repositories over time.