Allow for classmodel users to define which data should be collected
The radial bundle view uses language/classmodel to show the
project's namespaces, classes, and methods. For that purpose,
it requires the ability to disable the following classmodel
features:

1) The gathering of information about derived and base classes.
That would make radial bundle view enter an infinite loop.

2) The gathering of information about classes which aren't part
of the current opened project (the focus is only intra-project
dependencies).

3) The gathering of information about classes' internals (member
functions and attributes). One of the radial bundle view's aims is to
show dependencies between classes. Showing the methods for
big projects would make the visualization cluttered, so that we
need an option to not go into classes internals.