Do a bottom-up search when getting custom include paths.
When include paths are defined for the project folders

./        [1]
./foo     [2]
./foo/bar [3]

and we query for include paths defined for the file ./foo/bar/asdf.h,
then we need to iterate over the list of files such that we get
the include paths in the order of [3] + [2] + [1].

This patch fixes this and also amends the unit test.

Before, it behaved randomly due to QSet usage in KConfig::groupList.