Undeprecate the CvsJob functions.
We've lived with them for years, and they kept producing compiler warnings.
Since the functionality seems to work, let's undeprecate it. If someone
wants to fix this up and clean it up, please do. But I don't want to be spammed
with warnings in this old code that noone cares about anymore.