Revert "Use additional parser arguments for all files, not just for tests."
This reverts commit 9816f7b310d77ff75c91820929da0eb1a5a4e479.

This asserts due to access of ClangSettingsManager::parserSettings
from a background thread. Sergey, can you please look at that and
compile KDevelop in debug mode to see those assertions? As I said
in my email, I'm fine with this conceptually, but with this
assertion it's not a good idea to have it in.

CCMAIL: kalinichev.so.0@gmail.com