Revert "Revert "Revert "Use canonical path for symlinked subfolders in a project."""
OK, symlinks are clearly a mess. If we canonicalize paths this
early, then the folder names are unexpected (i.e. they use the
name of the target, not of the symlink).

Furthermore, quickopen etc. will only know the canonical path
and not the project local path, which is very bad.

I'll need to think about this some more to find a proper solution
for this ugly problem...

This reverts commit 6250737f544a3127a9f5620b436801aae3cf2d6a.