make qmljs kdevelop plugin compile on OSX
Summary:
It seems that when the qmljs plugin was ported over from qt-creator, the OSX specific files were left out.
 fileutils_mac.mm
 fileutils_mac.h
 autoreleasepool.h

Note: I simply copied these over from HEAD of qt-creator git.

Upon inspection of the license headers of other files in that directory, this appears to match what was done before.

This does include a modification to CMakeLists.txt to add the object-c++ file to compilation.

Test Plan: Build on OSX

Reviewers: kfunk, mwolff, apol

Reviewed By: apol

Subscribers: apol, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D456