Don't offer to implement functions related to Qt MOC.
The test is a bit crude so far, but far better than nothing.
We may get false-positives, if people hide functions named like
the moc functions in a macro that is exactly 8 chars wide now.

In the future, we could potentially improve this new heuristic by
checking whether the function is in a class context that inherits
QObject. I couldn't easily figure out how to do that right now
with the pure Clang API.