Fixup scope generation for nested classes in implementation helper.
This can be done by reusing a slightly adapted existing function
and throwing away quite some other code which is brittle. In the end
we now rely on clang_getCursorSemanticParent to build up the scope
chain for us which will work even when one of the parents is not
visited in the implementation helper.

While at it, I also fixed the handling of functions in anonymous
namespaces.

BUG: 355954