Keep the order of local declarations and child contexts on update.
This requires new API in DUContext but it fixes some serious issues
with kdev-clang, where on update we could end up with a broken order
of local declarations and/or child contexts in some conditions.

This then confuses other code build on top, such as the signature
assistant which adds a new test that gets fixed and another older
expected failure gets fixed as well.