Add new DUContext::resortLocalDeclarations and resortChildContexts API
This is required by kdev-clang and potentially other language plugins
that call setRange on a declaration or child context. In these cases
you can potentially break the order of the declarations and hell
breaks lose...

Now simply call one of the new functions after you are done with your
context, i.e. before closing it. See the upcoming kdev-clang patch
for more information.