OutputView: Remove output rewriting feature
Just show the plain output to the users, as he would see when invoking
the build tool inside the terminal.

Don't rewrite the output, always show the full output. I don't think
there's a reason to support rewriting the output altogether, it's only
misleading for the user.

As discussed in bug 354143

BUG: 354143
CCBUG: 194403
FIXED-IN: 5.0.0

Differential Revision: https://phabricator.kde.org/D757