Introduce IFilterStrategy::progressInLine
Makes it possible for filtering strategies to parse a line and report
progress if the line contains progress information.

Example make output:
  [  5%] Building foo
  [  6%] Building bar

A special make job filtering strategy can then parse the output and
report progress. The core idea is that this parsing is done in a
separate thread, like all the output filters do.

The progress information is also passed on to OutputExecuteJob which
emits the progress via KJob API.

Filter strategy implementations for Make & Ninja in kdevelop will follow.