Support of $VARIABLE in environment settings
BUG: 350183
FIXED-IN: 5.0.0

Summary:
Added the possibility to use $VARIABLE in environment settings.
After some discussion it was concluded, that we won't support defining a variable with other user defined variable. Only supported scenario is defining a variable with a system variable.

Test Plan: Tested with the added unit test and in practice.

Reviewers: kfunk

Subscribers: kdevelop-devel

Projects: #kdevelop

Differential Revision: https://phabricator.kde.org/D770