Always update the DUChain for the translation unit file.
This is required to properly update .cpp files to reflect changes
in header files. This patch is just a quick fix for the common case,
we might still run into issues for header files that depend on other
header files which got updated.

Also note that we need to introduce a central "DUChain::updateReady"
signal or similar, together with a mechanism to trigger that from
within the clang parse job. This is the reason why often the problem
view is not updating itself, leading to outdated highlighting of those.
Thing is: The problem view only listens to the background parser's
update signals, which do not get emitted for .cpp files when the .h
file is changed e.g. This is due to the different way the clang
lang plugin runs (single job can update multiple files), compared
to the oldcpp lang plugin (single job for single file).

BUG: 356327