Introduce a new DUChain::updateReady signal.
This is necessary to reliably track updates to the DUChain data.

The Clang parse job as a one-to-many relationship between the
parse job and to files it may trigger DUChain updates of.
Previously, a one-to-one relationship was always assumed, which
now uncovered various issues throughout KDevelop:

- assistants did not get properly updated
- problems and the highlighting of them did not update properly
- sometimes the context browser was also broken
- ...

Basically everywhere the code listened to finished ParseJobs to update
state when the DUChain was modified now broke apart. This new signal
simplifies code and fixes these issues.

Note that language plugins must call DUChain::emitUpdateReady now to
notify the world about updates. I tried to do this automatically
but we need a dedicated function as only language plugins know when
to emit this, i.e. when all processing on a given file has finished.

NOTE: THIS REQUIRES AN UPDATE TO ALL LANGUAGE PLUGINS!

CCBUG: 356327