progress widget: hide persistenly when completed
Nobody ever wants to open this widget and then have it pop up
again for every character typed. Plus, it can even be hard
to get rid of it because you only have a split-second to
click the hide button for small files.
Instead, once the current set of jobs is completed,
close the widget and only re-open it on request.