Show typename to the user even if duchain decl. wasn't found
Ideally we would always find a duchain declaration, but on
my very complex project with lots of preprocessor and
#include magic this often fails. The reason why it fails
is subject for future research. But in any case, we should
always show what we know to the user. Before this, many
objects were just annotated with the typename "<class>";
now it'll show the plain typename, similar to DelayedType.