Add manual test executable for the AbstractFileManagerPlugin import.
This can be used to benchmark the code by calling it with a path
to folders that contain lots of subfolders and files. Note that
even without the GUI latency issue introduced by QTBUG-50700 [1]
the code currently takes 10x as long as a recursive `find` call.
I will try to implement a non-blocking QDir based KIO alternative
for this purpose in the next days to speed up the common case of
loading a project from the local file system. This will then also
help as a workaround for the Qt bug mentioned above.

[1]: https://bugreports.qt.io/browse/QTBUG-50700