Workaround clang parsing problems when includes are missing
When an include was missing in a parse run, then the corresponding
include directive is removed from the parse using the unsaved-file
mechanism, and the document is re-parsed. This is repeated
until no more missing includes are found. The corresponding
missing-include problems are shown in the problem-reporter
anyway, so the user can fix it.
BUG: 358853