fix navigation in the context browser and navigation widget
When starting navigation at a member function, it wasn't
possible to take the 'back' link. There were probably also
other link instances that were broken. The problem was
that Qt5 seems to mess up the casing of the href link in
some cases, most obviosly it turns "A::b" into"a::b".
According to the HTML standard this is probably legal,
because some URI components are case-insensitive.
Therefore, simply create a new enumerated href ID for
each link.