fix navigation-widget usage in quickopen, code-completion and editor
Reminder for those who don't know the feature: The navigation-widget can
be activated using the keyboard by long-pressing the ALT key (it can
be navigated using ALT+arrow and ALT+return keys). This used
to work in the editor, in the code-completion list, and in the quickopen
list and outline. However, this was broken in KDev5 because the menu
bar steals the focus when the alt key is pressed without anything else
happening in between. The solution is to send some events to the
menu bar when ALT is needed for the navigation.