Also offer to implement function when in line of next function.
This fixes a bug where putting the cursor at the start of a line
where another function is defined, no implementations are offered.

This also workarounds the common case for bug 358521, where the
real cause is that the TU is not properly updated. Thus the
cursor will point to the old position where a function was and
then no offers where given.

CCBUG: 358521