duchain: builder: display offsets in bytes
Having the offsets of members in a structure displayed is extremely
useful for people who work with data structures. However the unit
chosen (bits) is not really common. This patch changes it to bytes. If
an offset doesn't round to a byte position, the extra bits are also
displayed.

REVIEW: 127061