Fix test failure uncovered by e90e1864039d8e.
Now that we always associate "file not found" errors with all files
parsed in a TU, we have to make sure all files are available. This
patch fixes that issue and ensure we actually test the right thing,
i.e. parse our overloaded qobjectdefs.h and not the "real" file.