Don't import the semantic parent if it equals the lexical parent.
We to import the semantic parent only if it differs from the lexical
parent, as then we deal with out-of-line declarations. For the other
case, the normal DUChain tree works fine. And importing the parent
again can lead to subtle differences, like DUChainUtils::getInheriters
misbehaving for nested classes.

This change should also have a nice positive performance impact.