Fixup colorization of class members for clang's DUChain structure.
The code in CodeHighlightingInstance::localClassFromCodeContext did
not work for the DUContext hierarchy build by kdev-clang: It puts
the function-body Other contexts into the Function context that
contains the arguments. This is very convenient, but was not supported
by the code that looks for the accompanying class during highlighting.

With this patch applied, we finally get the orangish colors for
members back! They no longer are colored like other globals!