Keep contexts alive when referenced by a problem.
This fixes the signature assistant when invoked from the header.
Now that we update the .cpp file when the header is changed, we
lost the declarations therein if e.g. the function return type got
changed. That then leads to issues in the signature assistant which
relies on the DUChain data.

I wonder whether we can find a more reliable way to store this data
to remove such kind of hacks... Problably we will have to work with
upstream clang and add FixIts for the signature assistant there
somehow...

CCBUG: 298511
CCBUG: 358479

Differential Revision: https://phabricator.kde.org/D1140