DUChain: Prioritize active documents
Prioritize document parsing in the background parser for documents which
just have been activated (i.e. the UI switched to a view of this
particular document).

Testing: Start up heaptrack session, with around 20 tabs open. Switch to
another tab at start up.
Observation: See that the current document in the tab gets a high priority in the background parser.

BUG: 358483
Differential Revision: D1145