Introduce/use KTextEditorHelpers::extractCursor
This is a copy of the command-line parsing code for filenames in the
kdevelop binary. It can be used to extract the line + column information
out of a string like "myfilename.cpp:1:2".

You can now also enter this in quick open, to open a file at a
particular line *and* column directly.