Refactor QtHelp plugin to provide a state.
The QtHelp plugin was conditionally created/deleted and hence a valid
object was taken as information if the qt documentation should be loaded.
After this refactoring, the plugin supports loading and unloading of the
documentation as well as information if a system QtHelp file could be
found. This is needed for status information in the GUI about existing
documentation files.

Performance-wise this change is not problematic, since this plugin is
only loaded for the configuration dialog.