Fix mouse usage inside the quickopen lists.
Don't hide the quickopen list on arbitrary WindowActivate or
WindowDeactivate events. Call 'checkFocus' more often,
and extend it so that it reliably fixes the visibility of
the quickopen list by checking the globally focussed widget
and window. Also make it call activateWindow in addition to
setFocus, so that it reliably puts the focus back into the
quick-open line edit when required.

BUG: 350409