Parse imports in the correct order.
This is critical for the correct resolution of declarations
and uses when headers influence each other.

Added two tests which failed before and succeed now. The tests
also test the more robust linking of uses which I committed
to kdevplatform, and which allows building functional uses
across headers that are dependent on each other but don't
include each other.