Fix possible crash in ProblemStore::addDiagnostic
The problem was, that child problem indices weren't
updated when the number of child problems didn't change.
When that happened, then child problem indices could
be wrong and basically random, thus creating a
recursive link which would normally be impossible.

The solution is, to always re-index the children
when storing the problem,  no matter whether
the child count changed or not.

BUG: 364971