If no pinned translation unit is available, ask the duchain
This fixes highlighting, navigation, etc. for complex cases
where the primitive "buddy" mechanism doesn't help finding
the translation unit. The buddy mechanism is not correct for
this case, because it was built around header/source pairs,
but fails in more complex cases. The duchain import chain
directly leads to the original transition unit from where
a header is included recursively.