PluginController::allPluginsForExtension: Return set
Make sure we return a unique set here, otherwise this may confuse the
documentation controller (and probably other places as well).

Scenario: Multiple plugins with the same plugin id in QT_PLUGIN_PATH
(for whatever reason). In this case PluginController::allPluginsForExtension would return
the same IPlugin instance for each of the available plugins => bogus.

This caused a crash on exit in the documentation component for me, since I
had multiple versions of kdevqthelp.so in QT_PLUGIN_PATH, and KDevelop
created bogus entries in the providers list of the documentation
controller. Crash is fixed with this patch.

BUG: 360405
FIXED-IN: 5.0.0