static assistants: rework problem display
The problem reporter plugin now fetches additional problems from the
static assistant manager directly, instead of adding them to the top
context. This removes the need for a write-lock in the foreground, and
also enables us to display squiggly problem lines directly when the
user is typing (not only after an update).