Hopefully fix screwed-up clang higlighting.
If you see this happening again, please tell me.
Make sure to also update kdevplatform.
CCMAIL:kdevelop-devel@kde.org
REVIEW:128480
The issue was that the clang plugin wouldn't store the modification
revision for files which were marked as "not modified" by KTextEditor.
Thus, no translation of the highlighting to the new state could be
performed for those documents. This is wrong, since a document can
be "not modified", but can still have changed in the time between the
start of the parse job start and end, etc. Just imagine you
press save after each character typed.