Do not perform highlighting when the revision is not held.
REVIEW:128480
This can happen when a new parse-job triggers and resets
the tracker in between of the end of a previous parse-job,
and the highlighter being run.

The proper fix would IMO be to attach the modification revision
reference to the respecive TopContext, and thus couple their
life times together. That would ensure we always have the revision
available as long as we have the top context (unlike the current
solution with storing the revision reference in the change tracker).