Make the selection of declarations under the cursor less annoying
Only include the back edge in matching when nothing could be found
without including the back edge. This makes it possible to reliably
point at uses of overloaded operators, and generally makes the
behavior more logical, because when you point at something, you
want that something to be selected, and not something else.