Avoid closing active tooltip inexpectedly
Sometimes the tooltip is shifted or reduced to a
geometry which isn't close enough to the mouse cursor,
so that it's not possible to reach the tooltip by
mouse without closing it. Always extend the tooltip
geometry by the position of the mouse cursor, so that
it's always possible to move the mouse cursor to the
tooltip without closing it.