Don't silently fail when trying to change write-protected files
Before, global find-replace and refactoring simply silently failed for
files which were missing write permissions on disk, and
document saving showed an error message without giving the user a chance
to fix the issue. Now a popup is shown with the list of affected files,
and the user is given an option to add write permissions to the files.