Keep patch-review tabs ordered correctly and enhance local patch support
After closing and re-opening files using the "Changes" list, tabs were
not ordered correctly like in the patch file.

Fixed the tab "Overview" title and other document properties of a
custom patch to be reset to default when the review has finished.

Also don't automatically update the patch when using a local patch file,
because the updating doesn't work properly for such files.