Use GDB as default debugger backend
Summary:
As discussed in IRC, make sure GDB is the default debugger backend when creating new launch profile.

Before this patch, it defaults to whatever comes first in the registered list.

This may not be the cleanest way to go. But adding new interfaces for setting default values from plugin side seems overkill.

Reviewers: kfunk

Reviewed By: kfunk

Subscribers: kdevelop-devel

Differential Revision: https://phabricator.kde.org/D2666