Fix PluginController::isEnabled
Make sure to respect user decisions about which plugin are supposed to
be loaded.

One problem to that: We can't change the 'enabled by default' property
'under the hood'. If the user loads a session, the plugin's activation
state is persisted in the session settings and preferred then. If we
were to change a plugin's 'enabled by default' property from false to
true; the user would still need to activate it himself. I don't think
that's an issue for us, though.

Differential Revision: D2618
BUG: 368264
FIXED-IN: 5.0.1