UI improvements
This commit adds ability to hiding (for context menu) and disabling (for main menu) cppcheck actions for following cases:

1) For non-project files actions are disabled.
2) Actions also disabled during running check.

This improves UI and remove showing of MessageBoxes (old code) when check is a-priori impossible.

Also this commit performs killing of running check if project is closed and clearing the ProblemView.