Do not use "script-error" icon for breakpoints
It just doesn't look right for some themes, e.g. Breeze. Instead, we
ship our own icon, which looks like what people expect (red dot).

At this occasion, use 16x16 size for the icon border icon,
because the width of the icon border is fixed to that in kate.

CCBUG:49719
Differential Revision: https://phabricator.kde.org/D3061