Fix Debuggers Actions Text
Summary: The patch fixes debuggers actions text when we have GDB and LLD plugins both enabled. This produce same text for both plugins and as a result we have 2 actions with text "Examine Core File..." and 2 with "Attach to Process...". This leads to user's impossibility to understand different between such actions.

Test Plan: Tested on mater branch

Reviewers: kfunk, #kdevelop

Reviewed By: kfunk, #kdevelop

Subscribers: kfunk, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D3122