Centralize code that deals with ideal tool view actions in one class
Summary:
Make it possible for to detect if the toolview widget is disabled

In such case, we show an asterisk by the button. We could do other
things such as changing the icon or such, but trying to minimize the noise.
Changing the tooltip accordingly could also make sense.

Test Plan: Manual testing, seems to just work

Reviewers: brauch, #kdevelop, kfunk

Reviewed By: brauch, #kdevelop, kfunk

Subscribers: kfunk, antonanikin, brauch, kdevelop-devel

Differential Revision: https://phabricator.kde.org/D3170