Fix extraction of multi-line UI strings from welcome page
xgettext was only extracting the first line of multi-line strings in QML.
The way to solve this is to use one string literal per line and concatenate
them with +, and also pass --language JavaScript to xgettext so it
understands the concatenation with +.

Unfortunately, xgettext can only handle one language at once. Fortunately,
this plugin doesn't have any i18n strings in the C++ code, so we can just
run xgettext in .qml files alone and skip the .cpp files.