Prefer qmake over qmake-qt5 over qmake-qt4.
In a previous patch of mine (8db2ca14) I preferred the versioned
qmake over the unversioned one, which was apparently required
back then on Arch. This is not the case anymore. The qmake in PATH
should be the right one to use, always.

If we don't find a fitting qmake, then we pick qmake-qt5 and only
then fall back to qmake-qt4.